Apply review feedbacks
