+++ etc/config/csharp.defaults.properties
@@ -10,7 +10,7 @@ compiler.csharp.clrDir=/opt/compiler-explorer/dotnet-6.0/artifacts/bin/coreclr/L
 compiler.csharp.runtimeId=linux-x64
 compiler.csharp.targetOS=linux
 compiler.csharp.targetArch=x64
-compiler.csharp.instructionSet=base,sse,sse2,sse3,ssse3,sse4.1,sse4.2,avx,avx2,aes,bmi,bmi2,fma,lzcnt,pclmul,popcnt,avxvnni
+compiler.csharp.targetInstructionSet=base,sse,sse2,sse3,ssse3,sse4.1,sse4.2,avx,avx2,aes,bmi,bmi2,fma,lzcnt,pclmul,popcnt,avxvnni
 compiler.csharp.targetFramework=net6.0
 compiler.csharp.buildConfig=Release
 compiler.csharp.jit=libclrjit.so
