+++ etc/config/fsharp.defaults.properties
@@ -10,7 +10,7 @@ compiler.fsharp.clrDir=/opt/compiler-explorer/dotnet-6.0/artifacts/bin/coreclr/L
 compiler.fsharp.runtimeId=linux-x64
 compiler.fsharp.targetOS=linux
 compiler.fsharp.targetArch=x64
-compiler.fsharp.instructionSet=base,sse,sse2,sse3,ssse3,sse4.1,sse4.2,avx,avx2,aes,bmi,bmi2,fma,lzcnt,pclmul,popcnt,avxvnni
+compiler.fsharp.targetInstructionSet=base,sse,sse2,sse3,ssse3,sse4.1,sse4.2,avx,avx2,aes,bmi,bmi2,fma,lzcnt,pclmul,popcnt,avxvnni
 compiler.fsharp.targetFramework=net6.0
 compiler.fsharp.buildConfig=Release
 compiler.fsharp.jit=libclrjit.so
