+++ etc/config/vb.defaults.properties
@@ -10,7 +10,7 @@ compiler.vb.clrDir=/opt/compiler-explorer/dotnet-6.0/artifacts/bin/coreclr/Linux
 compiler.vb.runtimeId=linux-x64
 compiler.vb.targetOS=linux
 compiler.vb.targetArch=x64
-compiler.vb.instructionSet=base,sse,sse2,sse3,ssse3,sse4.1,sse4.2,avx,avx2,aes,bmi,bmi2,fma,lzcnt,pclmul,popcnt,avxvnni
+compiler.vb.targetInstructionSet=base,sse,sse2,sse3,ssse3,sse4.1,sse4.2,avx,avx2,aes,bmi,bmi2,fma,lzcnt,pclmul,popcnt,avxvnni
 compiler.vb.targetFramework=net6.0
 compiler.vb.buildConfig=Release
 compiler.vb.jit=libclrjit.so
