+++ lib/compilers/vb.js
@@ -22,138 +22,8 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-import path from 'path';
-import fs from 'fs-extra';
+import { DotNetCompiler } from './dotnet';
 
-import { BaseCompiler } from '../base-compiler';
-
-export class VBCompiler extends BaseCompiler {
-
+export class VBCompiler extends DotNetCompiler {
     static get key() { return 'vb'; }
-
-    get rID() { return this.compilerProps('compiler.vb.runtimeId'); }
-    get targetFramework() { return this.compilerProps('compiler.vb.targetFramework'); }
-    get buildConfig() { return this.compilerProps('compiler.vb.buildConfig'); }
-
-    get compilerOptions() {
-        return ['publish', '-c', this.buildConfig, '--self-contained', '--runtime', this.rID];
-    }
-
-    async runCompiler(compiler, options, inputFileName, execOptions) {
-        if (!execOptions) {
-            execOptions = this.getDefaultExecOptions();
-        }
-
-        const programDir = path.dirname(inputFileName);
-        const clrBuildDir = this.compilerProps('compiler.vb.clrDir');
-
-        const projectFilePath = path.join(programDir, 'CompilerExplorer.vbproj');
-        const coreClrJitPath = path.join(clrBuildDir, this.compilerProps('compiler.vb.jit'));
-        const crossgen2Path = path.join(clrBuildDir, 'crossgen2', this.compilerProps('compiler.vb.crossgen2'));
-
-        const programPublishPath = path.join(
-            programDir,
-            'bin',
-            this.buildConfig,
-            this.targetFramework,
-            this.rID,
-            'publish',
-        );
-
-        const programDllPath = path.join(programPublishPath, 'CompilerExplorer.dll');
-
-        const projectFileContent =
-            `<Project Sdk="Microsoft.NET.Sdk">
-            <PropertyGroup>
-                <TargetFramework>${this.targetFramework}</TargetFramework>
-                <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-                <Nullable>enable</Nullable>
-                <AssemblyName>CompilerExplorer</AssemblyName>
-            </PropertyGroup>
-         </Project>
-        `;
-
-        execOptions.customCwd = programDir;
-        await fs.writeFile(projectFilePath, projectFileContent);
-        options = this.compilerOptions;
-
-        const compilerResult = await super.runCompiler(
-            compiler,
-            options,
-            inputFileName,
-            execOptions,
-        );
-
-        if (compilerResult.code !== 0)
-            return compilerResult;
-
-        const crossgen2Result = await this.runCrossgen2(
-            execOptions,
-            crossgen2Path,
-            coreClrJitPath,
-            programPublishPath,
-            programDllPath,
-            this.getOutputFilename(programDir, ''),
-        );
-
-        if (crossgen2Result.code !== 0)
-            return crossgen2Result;
-
-        return compilerResult;
-    }
-
-    optionsForFilter() {
-        return this.compilerOptions;
-    }
-
-    getOutputFilename(dirPath) {
-        return path.join(dirPath, `output.s`);
-    }
-
-    cleanAsm(asm) {
-        let cleanedAsm = '';
-        const asmLines = asm.replace(/\r/g, '').split('\n');
-
-        for (const line of asmLines) {
-            if (line.startsWith('; Assembly listing for method')) {
-                // ; Assembly listing for method ConsoleApplication.Program:Main(System.String[])
-                //                               ^ This character is the 31st character in this string.
-                // `substring` removes the first 30 characters from it and uses the rest as a label.
-                cleanedAsm = cleanedAsm.concat(line.substring(30) + ':\n');
-                continue;
-            }
-
-            if (line.startsWith(';') ||
-                line.startsWith('Emitting R2R PE file')) {
-                continue;
-            }
-
-            cleanedAsm = cleanedAsm.concat(line + '\n');
-        }
-
-        return cleanedAsm;
-    }
-
-    async runCrossgen2(execOptions, crossgen2Path, jitPath, publishPath, dllPath, outputPath) {
-        const targetOS = this.compilerProps('compiler.vb.targetOS');
-        const targetArch = this.compilerProps('compiler.vb.targetArch');
-        const instructionSet = this.compilerProps('compiler.vb.instructionSet');
-
-        const crossgen2Result = await this.exec(
-            crossgen2Path,
-            [
-                '--jitpath', jitPath, '-r', path.join(publishPath, '*'), dllPath, '-o', 'CompilerExplorer.r2r.dll',
-                '--targetos', targetOS, '--targetarch', targetArch, '--instruction-set', instructionSet,
-                '--codegenopt', 'NgenDisasm=*', '--codegenopt', 'JitDiffableDasm=1', '--parallelism', '1',
-            ],
-            execOptions,
-        );
-
-        await fs.writeFile(
-            outputPath,
-            this.cleanAsm(`${crossgen2Result.stdout}\n\n${crossgen2Result.stderr}`),
-        );
-
-        return crossgen2Result;
-    }
 }
